pyproject.toml
[project]
name = "rnd-strategy-agent"
version = "0.1.0"
description = "Agent-based R&D strategy analysis system for semiconductor technology trends and competitor intelligence"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain",
    "langgraph",
    "openai",
    "pydantic",
    "python-dotenv",
    "faiss-cpu",
    "pymupdf",
    "reportlab>=4.4.10",
    "tavily-python>=0.7.23",
]

[tool.uv]
dev-dependencies = [
    "ipykernel",
    "pytest",
    "ruff",
]
